Separate different rewrite behaviours into different classes
