+++ neptune-export/src/main/java/com/amazonaws/services/neptune/propertygraph/io/RewriteCsv.java
@@ -12,7 +12,6 @@ permissions and limitations under the License.
 
 package com.amazonaws.services.neptune.propertygraph.io;
 
-import com.amazonaws.services.neptune.io.Directories;
 import com.amazonaws.services.neptune.propertygraph.Label;
 import com.amazonaws.services.neptune.propertygraph.schema.*;
 import com.amazonaws.services.neptune.util.CheckedActivity;
@@ -22,10 +21,10 @@ import org.apache.commons.csv.CSVRecord;
 import org.apache.commons.lang3.ArrayUtils;
 
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
 import java.io.Reader;
-import java.util.*;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 public class RewriteCsv implements RewriteCommand {
@@ -51,16 +50,9 @@ public class RewriteCsv implements RewriteCommand {
         Map<Label, MasterLabelSchema> updatedSchemas = new HashMap<>();
 
         for (MasterLabelSchema masterLabelSchema : masterLabelSchemas.schemas()) {
-            if (targetConfig.mergeFiles()) {
-                updatedSchemas.put(
-                        masterLabelSchema.labelSchema().label(),
-                        rewriteAndMerge(targetConfig, graphElementType, masterLabelSchema));
-            } else {
-                updatedSchemas.put(
-                        masterLabelSchema.labelSchema().label(),
-                        rewrite(targetConfig, graphElementType, masterLabelSchema));
-            }
-
+            updatedSchemas.put(
+                    masterLabelSchema.labelSchema().label(),
+                    rewrite(targetConfig, graphElementType, masterLabelSchema));
         }
 
         return new MasterLabelSchemas(updatedSchemas, graphElementType);
@@ -137,126 +129,4 @@ public class RewriteCsv implements RewriteCommand {
         return masterLabelSchema;
     }
 
-    private MasterLabelSchema rewriteAndMerge(PropertyGraphTargetConfig targetConfig,
-                                              GraphElementType<?> graphElementType,
-                                              MasterLabelSchema masterLabelSchema) throws Exception {
-
-        LabelSchema masterSchema = masterLabelSchema.labelSchema();
-
-        String filename = Directories.fileName(String.format("%s.%s",
-                masterSchema.label().fullyQualifiedLabel(),
-                targetConfig.format().suffix()));
-
-        RenameableFiles renameableFiles = new RenameableFiles();
-
-        try (PropertyGraphPrinter printer = graphElementType.writerFactory().createPrinter(
-                filename,
-                masterSchema,
-                targetConfig.forFileConsolidation())) {
-
-            renameableFiles.add(new File(printer.outputId()), filename);
-
-            for (FileSpecificLabelSchema fileSpecificLabelSchema : masterLabelSchema.fileSpecificLabelSchemas()) {
-
-                try (DeletableFile file = new DeletableFile(new File(fileSpecificLabelSchema.outputId()))) {
-
-                    LabelSchema labelSchema = fileSpecificLabelSchema.labelSchema();
-                    Label label = labelSchema.label();
-
-                    String[] additionalElementHeaders = label.hasFromAndToLabels() ?
-                            new String[]{"~fromLabels", "~toLabels"} :
-                            new String[]{};
-
-                    String[] filePropertyHeaders =
-                            labelSchema.propertySchemas().stream()
-                                    .map(p -> p.property().toString())
-                                    .collect(Collectors.toList())
-                                    .toArray(new String[]{});
-
-                    String[] fileHeaders = ArrayUtils.addAll(
-                            graphElementType.tokenNames().toArray(new String[]{}),
-                            ArrayUtils.addAll(additionalElementHeaders, filePropertyHeaders));
-
-                    try (Reader in = file.reader()) {
-
-                        CSVFormat format = CSVFormat.RFC4180.withHeader(fileHeaders);
-                        Iterable<CSVRecord> records = format.parse(in);
-
-                        for (CSVRecord record : records) {
-                            printer.printStartRow();
-
-                            if (graphElementType.equals(GraphElementTypes.Nodes)) {
-                                printer.printNode(record.get("~id"), Arrays.asList(record.get("~label").split(";")));
-                            } else {
-                                if (label.hasFromAndToLabels()) {
-                                    printer.printEdge(
-                                            record.get("~id"),
-                                            record.get("~label"),
-                                            record.get("~from"),
-                                            record.get("~to"),
-                                            Arrays.asList(record.get("~fromLabels").split(";")),
-                                            Arrays.asList(record.get("~toLabels").split(";")));
-                                } else {
-                                    printer.printEdge(record.get("~id"), record.get("~label"), record.get("~from"), record.get("~to"));
-                                }
-                            }
-
-                            printer.printProperties(record.toMap(), false);
-                            printer.printEndRow();
-                        }
-
-                    }
-                }
-            }
-
-        }
-
-        renameableFiles.rename();
-
-        return new MasterLabelSchema(
-                masterSchema,
-                Collections.singletonList(new FileSpecificLabelSchema(filename, targetConfig.format(), masterSchema)));
-    }
-
-    private static class RenameableFiles {
-
-        private final Map<File, String> entries = new HashMap<>();
-
-        public void add(File file, String filename) {
-            entries.put(file, filename);
-        }
-
-        public void rename() {
-            for (Map.Entry<File, String> entry : entries.entrySet()) {
-                File file = entry.getKey();
-                file.renameTo(new File(file.getParentFile(), entry.getValue()));
-            }
-        }
-    }
-
-    private static class DeletableFile implements AutoCloseable {
-
-        private final File file;
-
-        private DeletableFile(File file) {
-            this.file = file;
-        }
-
-        public Reader reader() throws FileNotFoundException {
-            return new FileReader(file);
-        }
-
-        public String name() {
-            return file.getName();
-        }
-
-        @Override
-        public void close() {
-            boolean deletedOriginalFile = file.delete();
-
-            if (!deletedOriginalFile) {
-                throw new IllegalStateException("Unable to delete file: " + file.getAbsolutePath());
-            }
-        }
-    }
 }
